requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
opencv-python-headless